Integrate Level manager
Adds Level manager as a part of Elmanager. It can be launched from the main menu or directly from the exe file.

The Level manager project has been dormant since August 2009, so the code quality of the project is not terribly high as I had just started learning PC programming back then. But since it nonetheless works fine, it's useful and I decided to add it.

The source code files in LevelManager folder are all original (from August 2009) except Main.vb which I added for better compatibility with Elmanager.

Closes #78.