Define custom shapes (#158)
This pull request adds a new Custom Shape Tool, called "Shape", to the
Level Editor that allows you to quickly place predefined custom shapes
into the current level.
Furthermore, this pull request also adds a new right click context menu
option called "Save as shape..." which creates a new custom shape based
on the current selection.

Custom shapes are saved as regular Elma levels (*.lev) on disk, and are
grouped by sub-folders in the root folder for custom shapes. The root
folder for custom shapes is called `sle_shapes` and will be created in
the same directory as Elmanager.exe the first time a custom shape is
saved.

The custom shape collection is just a folder structure with files, and
therefore easily shareable / importable / exportable.

The following components can be added to custom shapes:
- Polygons (always full polygons, if any vertex is selected on a
polygon, the whole polygon is serialized).
 - Grass polygons
 - Objects (Apples, Killers, Flowers)
 - Pictures
 - Textures

The new Custom Shape Tool supports the following actions:
 - Right click to open the Shape Selection dialog
 - Anchoring using num 1-5 like Picture Tool.
 - Transformation
   - Mirroring - Vertical, Horizontal or Both (toggle with num 6)
- Rotation - num 7 to rotate left, num 9 to rotate right, num 8 to reset
rotation
   - Scaling - Resize using (+ / - ) on the keyboard
   - Reset anchoring and transformations: num 0
- The transformation state is remembered between shapes, so if you
rotate one shape, and select a different one, the same settings apply to
the new shape as well.

# Images
### New Shape button, and Shape Selection dialog

![Elmanager_6LHlIvoMYi](https://github.com/user-attachments/assets/1e297d05-39b6-4046-8d23-6adb9c14d818)


### Shapes are grouped by folders.

![Elmanager_ckvefUHXjf](https://github.com/user-attachments/assets/85795a18-a164-4269-a4cd-2523d8627456)



# Example shapes for quick testing
Here is an example collection of shapes you can use for trying it out.
Place the following `sle_shapes` folder in the same folder as the
Elmanager.exe

[sle_shapes.zip](https://github.com/user-attachments/files/19263475/sle_shapes.zip)

---------

Co-authored-by: Mika Lehtinen <Smibu@users.noreply.github.com>
Co-authored-by: Smibu <smibu90@gmail.com>